+++ libs/core/iostream/tests/unit/detail/write_seekable_seq_test.hpp
@@ -15,9 +15,9 @@
 #include <hpx/modules/iterator_support.hpp>
 #include <hpx/modules/testing.hpp>
 
+#include <cstddef>
 #include <fstream>
 #include <vector>
-
 #include "sequence.hpp"
 #include "temp_file.hpp"
 #include "verification.hpp"
@@ -32,7 +32,8 @@ void write_seekable_sequence_test()
     test_file test;
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         filtering_stream<seekable> out(
             hpx::util::iterator_range(first.begin(), first.end()), 0);
         write_data_in_chars(out);
@@ -43,7 +44,8 @@ void write_seekable_sequence_test()
     }
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         filtering_stream<seekable> out(
             hpx::util::iterator_range(first.begin(), first.end()), 0);
         write_data_in_chunks(out);
@@ -54,7 +56,8 @@ void write_seekable_sequence_test()
     }
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         filtering_stream<seekable> out(
             hpx::util::iterator_range(first.begin(), first.end()));
         write_data_in_chars(out);
@@ -65,7 +68,8 @@ void write_seekable_sequence_test()
     }
 
     {
-        vector<char> first(data_reps * data_length(), '?');
+        vector<char> first(
+            static_cast<std::size_t>(data_reps * data_length()), '?');
         filtering_stream<seekable> out(
             hpx::util::iterator_range(first.begin(), first.end()));
         write_data_in_chunks(out);
